Move tests to test/macaroni dir
